move GraphQL queries to their own directory
