+++ data/queries/pull_request.graphql
@@ -1,36 +1,3 @@
-IssueQuery: str = """
-{
-  author {
-    login
-    url
-    avatarUrl
-  }
-  url
-  createdAt
-  closed
-  closedAt
-  bodyText
-  title 
-  number
-  state
-  comments {
-    totalCount
-  }
-  participants {
-    totalCount
-  }
-  assignees {
-    totalCount
-  }
-  labels(first: 100) {
-    nodes {
-      name
-    }
-  }
-}
-"""
-
-PullRequestQuery: str = """
 {
   title
   url
@@ -105,4 +72,3 @@ PullRequestQuery: str = """
     }
   }
 }
-"""
